# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EpiCare AI",
    "description": "\n**EpiCare AI** is a robust and comprehensive platform designed to manage and analyze epilepsy-related data, facilitating seamless interactions between patients, doctors, and caretakers.\n\n### Core Features\n- **Admin Dashboard**: Comprehensive system metrics and doctor PMDC verification tools.\n- **User Management**: Unified profiles tailored for Patients, Doctors, and Caretakers.\n- **Role-Based Authentication**: Secure access control with PMDC verification for doctors.\n- **Connection System**: Connect patients with verified medical professionals and trusted caretakers.\n- **AI-Driven EEG Analysis**: Upload EEG sessions, process signals, and receive detailed AI-generated seizure prediction reports.\n- **Personalized Dashboard**: Get real-time heuristic recommendations based on daily habits.\n- **Lifestyle & Medication Tracking**: Log sleep, stress, triggers, and medications with adherence algorithms.\n- **Emergency SOS**: Instantly dispatch multi-channel (Firebase/WhatsApp/Email) SOS alerts to configured contacts.\n\nThis API adheres strictly to REST principles, delivering standardized responses and explicit HTTP status codes to ensure a reliable developer experience.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/seizures/manual": {
      "get": {
        "tags": [
          "Manual Seizure Logs"
        ],
        "summary": "List Manual Seizures",
        "description": "Retrieve a descending history of all manually logged seizures for the authenticated patient.",
        "operationId": "get_manual_seizures_api_v1_seizures_manual_get",
        "responses": {
          "200": {
            "description": "A list of manual seizure logs.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ManualSeizureLogOut"
                  },
                  "type": "array",
                  "title": "Response Get Manual Seizures Api V1 Seizures Manual Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Manual Seizure Logs"
        ],
        "summary": "Log Manual Seizure",
        "description": "Manually log a seizure that occurred while the patient was not wearing an EEG device. Tracks core clinical metrics like duration, auras, and post-ictal symptoms.",
        "operationId": "log_manual_seizure_api_v1_seizures_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualSeizureLogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The newly created manual seizure log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualSeizureLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/seizures/manual/{log_id}": {
      "put": {
        "tags": [
          "Manual Seizure Logs"
        ],
        "summary": "Update Manual Seizure Log",
        "description": "Updates a manual seizure log to correct accidental mistakes.",
        "operationId": "update_manual_seizure_api_v1_seizures_manual__log_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Log Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualSeizureLogUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualSeizureLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Manual Seizure Logs"
        ],
        "summary": "Delete Manual Seizure Log",
        "description": "Deletes a manual seizure log completely.",
        "operationId": "delete_manual_seizure_api_v1_seizures_manual__log_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualSeizureLogCreate": {
        "properties": {
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "seizure_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seizure Type"
          },
          "auras_felt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auras Felt"
          },
          "post_ictal_symptoms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Ictal Symptoms"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "occurred_at",
          "duration_seconds"
        ],
        "title": "ManualSeizureLogCreate"
      },
      "ManualSeizureLogOut": {
        "properties": {
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "seizure_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seizure Type"
          },
          "auras_felt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auras Felt"
          },
          "post_ictal_symptoms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Ictal Symptoms"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "occurred_at",
          "duration_seconds",
          "id",
          "user_id"
        ],
        "title": "ManualSeizureLogOut"
      },
      "ManualSeizureLogUpdate": {
        "properties": {
          "occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "seizure_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seizure Type"
          },
          "auras_felt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auras Felt"
          },
          "post_ictal_symptoms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Ictal Symptoms"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ManualSeizureLogUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}